front_back_interaction/webrtc/back/pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["Nikita"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-socketio = "^5.11.3"
uvicorn = "^0.30.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
